Move PersistAllocatedLicenceToResultsService
